♻️ Refactor forecasting routes and models into the package
